+++ rust/otel-arrow-rust/src/lib.rs
@@ -18,7 +18,7 @@
 #[allow(dead_code)]
 pub(crate) mod arrays;
 
-pub mod decode;
+mod decode;
 mod error;
 pub mod otap;
 pub mod otlp;
